# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Reserva de Laboratórios e Salas",
    "version": "1.0.0",
    "description": "Importe no Insomnia. Cadastre usuário, confira login e configure Basic Auth. Use IDs retornados nas consultas. Datas DD/MM/AAAA e horários HH:MM:SS. Testes ainda pendentes."
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "security": [
    {
      "basicAuth": []
    }
  ],
  "paths": {
...
    "/api/salas": {
      "get": {
        "summary": "Consultar salas",
        "tags": [
          "salas"
        ],
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RecursoView"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Autenticação inválida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Registro não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Conflito",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Erro de validação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "nome",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "capacidade",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "example": 1
            }
          },
          {
            "name": "localizacao",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "statusRecursoCodigo",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "example": 1
            }
          },
          {
            "name": "data",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^\\d{2}/\\d{2}/\\d{4}$",
              "example": "30/09/2026",
              "description": "DD/MM/AAAA; use uma data futura na reserva"
            }
          },
          {
            "name": "hora",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^\\d{2}:\\d{2}:\\d{2}$",
              "example": "10:00:00"
            }
          }
        ]
      },
      "post": {
        "summary": "Cadastrar salas",
        "tags": [
          "salas"
        ],
        "responses": {
          "201": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecursoView"
                }
              }
            },
            "headers": {
              "Location": {
                "description": "URI do registro criado",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Autenticação inválida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Registro não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Conflito",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Erro de validação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecursoInput"
              }
            }
          }
        }
      }
    },
    "/api/salas/{codigo}": {
      "get": {
        "summary": "Buscar por código",
        "tags": [
          "salas"
        ],
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecursoView"
                }
              }
            }
          },
          "401": {
            "description": "Autenticação inválida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Registro não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Conflito",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Erro de validação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "codigo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "example": 1
            }
          }
        ]
      },
      "put": {
        "summary": "Atualizar cadastro completo",
        "tags": [
          "salas"
        ],
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecursoView"
                }
              }
            }
          },
          "401": {
            "description": "Autenticação inválida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Registro não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Conflito",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Erro de validação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecursoInput"
              }
            }
          }
        },
        "parameters": [
          {
            "name": "codigo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "example": 1
            }
          }
        ]
      },
      "delete": {
        "summary": "Excluir registro (reserva ativa respeita prazo de cancelamento)",
        "tags": [
          "salas"
        ],
        "responses": {
          "204": {
            "description": "Sucesso"
          },
          "401": {
            "description": "Autenticação inválida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Registro não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Conflito",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Erro de validação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "codigo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "example": 1
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic"
      }
    },
    "schemas": {
...
      "StatusView": {
        "type": "object",
        "properties": {
          "codigo": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "nome": {
            "type": "string"
          }
        },
        "required": [
          "codigo",
          "nome"
        ],
        "additionalProperties": false
      },
      "RecursoInput": {
        "type": "object",
        "properties": {
          "nome": {
            "type": "string",
            "minLength": 3,
            "maxLength": 80,
            "example": "Laboratório de Informática"
          },
          "capacidade": {
            "type": "integer",
            "minimum": 1,
            "maximum": 40,
            "example": 30
          },
          "localizacao": {
            "type": "string",
            "minLength": 15,
            "maxLength": 50,
            "example": "Bloco A - Primeiro andar"
          },
          "statusRecursoCodigo": {
            "type": "integer",
            "format": "int64",
            "example": 1
          }
        },
        "required": [
          "nome",
          "capacidade",
          "localizacao",
          "statusRecursoCodigo"
        ],
        "additionalProperties": false
      },
      "RecursoView": {
        "type": "object",
        "properties": {
          "codigo": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "tipo": {
            "type": "string",
            "enum": [
              "LABORATORIO",
              "SALA"
            ]
          },
          "nome": {
            "type": "string"
          },
          "capacidade": {
            "type": "integer"
          },
          "localizacao": {
            "type": "string"
          },
          "statusRecurso": {
            "$ref": "#/components/schemas/StatusView"
          }
        },
        "required": [
          "codigo",
          "tipo",
          "nome",
          "capacidade",
          "localizacao",
          "statusRecurso"
        ],
        "additionalProperties": false
      },
...
      "ApiError": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "integer"
          },
          "mensagem": {
            "type": "string"
          },
          "caminho": {
            "type": "string"
          },
          "erros": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "campo": {
                  "type": "string"
                },
                "mensagem": {
                  "type": "string"
                }
              },
              "required": [
                "campo",
                "mensagem"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "timestamp",
          "status",
          "mensagem",
          "caminho",
          "erros"
        ],
        "additionalProperties": false
      }
    }
  }
}